# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCE API",
    "description": "Code context engine: indexing, search, relations and tool endpoints",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/project/{project_id}/call-chain/{id}": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "Handle call chain request",
        "operationId": "handle_call_chain",
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "max_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "exclude_tests",
            "in": "query",
            "description": "Exclude entities located in test files",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "directory_prefix",
            "in": "query",
            "description": "Keep only entities under this directory prefix",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excluded_files",
            "in": "query",
            "description": "Exact file paths to exclude",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Entity id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallChainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Index unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{project_id}/call-path": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "Handle call path request",
        "operationId": "handle_call_path",
        "parameters": [
          {
            "name": "start_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "max_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallPathResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Index unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{project_id}/class/{id}/implementations": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "Handle class implementations request",
        "operationId": "handle_class_implementations",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Class id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassImplementationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Index unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{project_id}/class/{id}/inheritance": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "Handle class inheritance request",
        "operationId": "handle_class_inheritance",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Class id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassInheritanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Index unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CallChainResponse": {
        "type": "object",
        "description": "Call chain response",
        "required": [
          "success",
          "relation_epoch",
          "function_id",
          "function_name",
          "direction",
          "call_chain"
        ],
        "properties": {
          "call_chain": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CallChainNode"
            }
          },
          "direction": {
            "type": "string"
          },
          "function_id": {
            "type": "string"
          },
          "function_name": {
            "type": "string"
          },
          "relation_epoch": {
            "type": "integer",
            "format": "int64"
          },
          "relation_info": {},
          "success": {
            "type": "boolean"
          }
        }
      },
      "CallPathResponse": {
        "type": "object",
        "description": "Call path response",
        "required": [
          "success",
          "relation_epoch",
          "start_function_id",
          "end_function_id",
          "path_found",
          "path_length"
        ],
        "properties": {
          "end_function_id": {
            "type": "string"
          },
          "path": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CallChainNode"
            }
          },
          "path_found": {
            "type": "boolean"
          },
          "path_length": {
            "type": "integer",
            "minimum": 0
          },
          "relation_epoch": {
            "type": "integer",
            "format": "int64"
          },
          "relation_info": {},
          "start_function_id": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ClassImplementationsResponse": {
        "type": "object",
        "description": "Class implementations response",
        "required": [
          "success",
          "relation_epoch",
          "class_id",
          "class_name",
          "implemented_interfaces",
          "implementing_classes"
        ],
        "properties": {
          "class_id": {
            "type": "string"
          },
          "class_name": {
            "type": "string"
          },
          "implemented_interfaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterfaceRelation"
            }
          },
          "implementing_classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassRelation"
            }
          },
          "relation_epoch": {
            "type": "integer",
            "format": "int64"
          },
          "relation_info": {},
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Common error response",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}